Added generating of thing parameter configuration file. (#4771)
To get access to thing configuration parameters a class can be used. Very often new bindings don't know this feature and use the properties class directly and implement their own parsing of the parameters. By using a configuration class this is all not needed. This change adds a sample configuration file so people directly see what is possible.

Signed-off-by: Hilbrand Bouwkamp <hilbrand@h72.nl>