move generic UID stuff to a seperate class
Currently the UID is part of the core.thing bundle and states that every
UID must start with the binding ID.
To reuse the non-binding, non-thing related UID implementation a new
base class for a generic UID has been created.
It is part of the core bundle. So no need to depend on the thing bundle
or to start the UID with a binding ID.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>