[5198] prevent Responses to be interpreted as xml
The ResponseBuilder "guesses" the best mimetype for the Response by
interpreting the accept headers from the Request. When the client
sets no preferation and xml is valid, the response will be interpreted
as xml. If FF doesn't find the xml root element in this case, which happens
when the Response is empty, a console error is printed out.

Bug: https://github.com/eclipse/smarthome/issues/5198
Signed-off-by: Marcel Walwei <maw.albrecht.jung@gmail.com>