fix reference mess when updating uninitialized things (#5234)
Whenever there was a ThingHandler which failed to initialize,
e.g. because of a missing/non-configured bridge, then the framework
would not updated this ThingHandler's Thing reference as soon
as there was an update on that Thing. As a result, the ThingHandler
operated on the old Thing instance with all its consequences.

fixes #5215
Signed-off-by: Simon Kaufmann <simon.kfm@googlemail.com>