CronExpression: start using fixed test input
The current test execution depends on the TimeZone and current time of
the system the test is executed. That is not optimal as tests could
succeed on the one machine but fail on the other one.
Added another test (currently ignored, too), that uses a explicit set
TimeZone and time.
We use "millis from Epoc" to set the time, so we can easily difference
between e.g. 2:30 nDST and 2:30 DST.

Related to: https://github.com/eclipse/smarthome/issues/5294
Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>