Item tags become metadata
With this change, the tags which previously were stored
together with the items now are stored as metadata. This
is completely transparent though, so that users won't
require to migrate any data.

They are, however, stored in a namespace (`_tags`)  which
by its prefix marks it as "internal" and therefore is not
exposed like all other, normal namespaces e.g. through the
REST API.

fixes #5428
Signed-off-by: Simon Kaufmann <simon.kfm@googlemail.com>