Bindings Archetype: Move Handler and Constants into internal package
As discussed in #4330 it makes sense to export handlers to create a derived
handler in another bundle. We do have this for the bluetooth binding and
will have it for the zigbee binding as well.

These two bindings are the rare exception and their handler APIs have been
throught through, i.e. they have a stable API and do not expose internal
classes.

Thus for newly created bindings it makes sense to NOT export the handlers
and if there should be a derived handler at some point in the future we
know the use-cases for the derived handler and can design a proper API
for the handler at that time and then export the handler.

Fixes #4330

Signed-off-by: Stefan Triller <stefan.triller@telekom.de>