fix Item confusion (#5607)
* fix Item confusion

In #5207 the AbstractRegistry was changed to call
onUpdateElement(...) with the current live instance
instead of the oldElement instance which got handed
in by the provider.

However, the given oldElement was guaranteed to contain
the previous values, while the live instance potentially
is subject to modifications. At least as long as there
are e.g. ActiveItem & GenericItem part of the public API
and allow modification of the entities.

Signed-off-by: Simon Kaufmann <simon.kfm@googlemail.com>