Change unit delimiter from [ ] to |
...as the previous solution was conflicting with the XClosure rule.

The new syntax uses the pipe symbol as a delimiter. As long as the
unit symbols match the ID rule, they can be written as is, otherwise
they will have to noted as strings. In Script.xtext, some symbols are
included as keywords in order to facilitate the most commonly used
ones. For everything else, strings must be used.

The notation therefore looks like the following examples, whereby all
of them evaluate to the same QuantityType:

   20|°C
   20 | °C
   "20|"°C"
   20 | "°C"

The XBase language doesn't define bitwise operators, therefore I currently
do not see any potential conflics with other rules. The pipe symbol is used
in XClosure, but doesn't match the preceeding Number. Therefore this should
be okay.

Signed-off-by: Simon Kaufmann <simon.kfm@googlemail.com>