[Tradfri] further improve connection handling (#5631)
This change continues #5623 by doing some small additional
improvements in order to let the GC do its job.

Furthermore, with #5623 in place, the restart of the transport
layer as done in #3675 turned out to be not needed anymore.
Therefore it is reverted in this PR again. This allows the
reliability layer to do its job and e.g. re-deliver commands
within a certain timeframe in case of a short connection loss.

relates to #5622
Signed-off-by: Simon Kaufmann <simon.kfm@googlemail.com>