POM Code Convention
I realized that there is at least one pom.xml file that uses tabs
instead of spaces for indentation. So I started the POM cleanup again.

In #5338 I missed to set the predefined sort order profile.
IMHO we should use the "recommended_2008_06" profile that states:
The POM Code Convention that was chosen by Maven developers in 2008

The main difference to the current one and the reason so much files are
touched is that the order for GAV has been "corrected".
In the default profile the artifactId is above the groupId (which does
not make sense IMHO at all).

Sorry for not using the correct profile the first time.

Command that has been executed:

    mvn \
      com.github.ekryd.sortpom:sortpom-maven-plugin:sort \
      -Dsort.keepBlankLines=true \
      -Dsort.predefinedSortOrder=recommended_2008_06

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>