JSONStorage: Do not use invalid fileNames for some file-systems
Windows has a rather long list of forbidden characters in filenames, see

https://msdn.microsoft.com/en-us/library/aa365247(VS.85).aspx

Linux-like systems only disallow / and the null-byte.

https://en.wikipedia.org/wiki/Filename#Reserved_characters_and_words

By URLencoding the string, all of these characters are safely transformed
into %<HexValue> and % is an allowed character. This encoding is even
reversible.

Given that we will not run on RT-11 systems we should be safe.

I use a maximum length of 127 to be safe even on android systems:
https://stackoverflow.com/questions/13204807/max-file-name-length-in-android

The down-side is that even spaces are now converted into %20, but I chose
to use URLencoder.encode instead of implementing the conversion by myself.

For existing storages I check whether a file exists and keep using it.

Fixes #5687

Signed-off-by: Stefan Triller <stefan.triller@telekom.de>