[Homematic] Provide duty cycle channel for homematic bridges (#5821)
- Introduces a new channel on the homematic bridge type called
DUTY_CYCLE_RATIO. The channel shows the current duty cycle usage of the
bridge in percent. The bridge is polled for this value every 15 seconds.

- When the bridge reports a duty cycle of 99% or more, the bridge goes
into the status "OFFLINE/DUTY_CYCLE" and goes back online once it drops
below this threshold.

This will make it easier to undersand for the user when the homematic
installation does not work due to the duty cycle.

Signed-off-by: Florian Stolte <fstolte@itemis.de>