add nullness annotations to JavaTest
There is a waitForAssert method that check if the given runnable
"assertion" is null on some place. So it seems it supports the usage of
a nullable reference. But later `assertion.run()` is called without
handling a null reference. I cannot find any LOC that uses a null
reference and so I leave that member non-null.

I identified a method used by the firmware tests that intention is only
to return null to workaround the nullness checks at compile time. I
improved the JavaDoc to explain its intention more detailed. I am not
sure if we should provide such a method at all, but I assume for testing
purposes it could perhaps make sense. So I leave it as it is. If we want
to use another logic and remove that method, it could be done in a
separate commit / PR later.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>