add nullness annotations to JavaOSGiTest
Before the method that returns the first found implemented interface of
an object returns null if no interface is implemented. The return value
is taken and supplied to methods that does require a non-null interface
name. Instead of adding checks and immediate returns on every usage, I
changed the method to throw an IAE if an object is given that does not
implement an interface. As it has been worked before it is currently not
used on objects that does not implement an interface. And as it is
intended to be used for register and unregister services it does not
make sense to be used on such instances at all.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>