fix unregister thing handler factory of thing registry OSGi test
The test uses an anonymous inner class instance of a thing handler
factory. That instance extends the BaseThingHandlerFactory.
It register that instance as a thing handler factory service and tries
to unregister that service later using
`unregisterService(thingHandlerFactory)`.
The unregisterService method with an object reference takes the first
interface name returned by the `getInterfaceName` method and unregister
that service.
The `getInterfaceName` only inspect the directly implemented interfaces
but not the interfaces that are implemented by e.g. parent class.
The usage of an anonymous inner class that only extends a class that
implements that interface but does not declare that interface
implementation itself (it is an anonymous inner class) no interface name
could be found and no service it deregistered at all.
The logic has been changed to safe the registered service reference and
uses that one to unregister the service correctly.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>