add nullness annotations to JavaTest, JavaOSGiTest (#5838)
* add nullness annotations to JavaTest

There is a waitForAssert method that check if the given runnable
"assertion" is null on some place. So it seems it supports the usage of
a nullable reference. But later `assertion.run()` is called without
handling a null reference. I cannot find any LOC that uses a null
reference and so I leave that member non-null.

I identified a method used by the firmware tests that intention is only
to return null to workaround the nullness checks at compile time. I
improved the JavaDoc to explain its intention more detailed. I am not
sure if we should provide such a method at all, but I assume for testing
purposes it could perhaps make sense. So I leave it as it is. If we want
to use another logic and remove that method, it could be done in a
separate commit / PR later.

* add nullness annotations to JavaOSGiTest

Before the method that returns the first found implemented interface of
an object returns null if no interface is implemented. The return value
is taken and supplied to methods that does require a non-null interface
name. Instead of adding checks and immediate returns on every usage, I
changed the method to throw an IAE if an object is given that does not
implement an interface. As it has been worked before it is currently not
used on objects that does not implement an interface. And as it is
intended to be used for register and unregister services it does not
make sense to be used on such instances at all.

* fix unregister thing handler factory of thing registry OSGi test

The test uses an anonymous inner class instance of a thing handler
factory. That instance extends the BaseThingHandlerFactory.
It register that instance as a thing handler factory service and tries
to unregister that service later using
`unregisterService(thingHandlerFactory)`.
The unregisterService method with an object reference takes the first
interface name returned by the `getInterfaceName` method and unregister
that service.
The `getInterfaceName` only inspect the directly implemented interfaces
but not the interfaces that are implemented by e.g. parent class.
The usage of an anonymous inner class that only extends a class that
implements that interface but does not declare that interface
implementation itself (it is an anonymous inner class) no interface name
could be found and no service it deregistered at all.
The logic has been changed to safe the registered service reference and
uses that one to unregister the service correctly.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>