WIP: Add tags to SystemChannelTypes
This PR is meant for discussing the tags for the default SystemChannelTypes
as stated in the wiki:

https://github.com/eclipse/smarthome/wiki/Semantic-Tag-Library

Also the location (and name) for the "SystemTags" class is open for
discussion.

The idea is to have hierarchical tags, so "Purpose:Alarm:Fire" can also
be retrieved by looking for all alarms, i.e "Purpose:Alarm".

Once we have agreed on the tags, I will adjust the tests for the channels
and add them to this PR.

Signed-off-by: Stefan Triller <stefan.triller@telekom.de>