Automation: I/O connection supports HashMaps
Automation modules support input and output values and their names and/or
types have to match in order to be a valid connection between two modules.

With this PR it is possible that one module returns a HashMap of values
and a follow-up modul fetches certain values out of this map. This allows
for creating modules with a dynamic amount of outputs.

For example an action reference the values of a trigger's output called
"itemStates" as follows:

```
"inputs" : {
	"stateItemA" : "MyTrigger.itemStates.ItemA",
	"stateItemB" : "MyTrigger.itemStates.ItemB"
}
```

Where "itemStates" is the HashMap and "ItemA" and "ItemB" are the keys.
Input variable "stateItemA" will be whichever value is the result of the
call `itemStates.get("ItemA")`.

Signed-off-by: Stefan Triller <stefan.triller@telekom.de>