thing status details as in documentation
Sometimes I need to open the documentation to identify the correct
status details for a specific thing status.

There exist currently two specific status detail classes (offline and
online) but we miss specific classes for uninitialized and other states.

This adds classes and singleton instances of possible status details for
every status.

It also adds the "GONE" detail to the offline status as documented.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>