[homematic] Fix issue 6053: HmIP devices offline
With this fix, the RpcClient does not try to retrieve information about
the VALUES Paramset of a logical device anymore, if that logical device
does not correspond to a channel.

This is necessary since there is a non-channel logical device for each
physical device, and this logical device only has a MASTER Paramset.

When the RpcClient tried to retrieve the VALUES Paramset for a
non-channel logical device, it received an Error in case of HmIP, which
results in the device not being successfully discovered by the binding.

Signed-off-by: Florian Stolte <fstolte@itemis.de>