drop tycho compiler warning about missing project specific JDT settings
The Maven build currently shows a warning for every project that looks
like:

    [WARNING] Parameter 'useProjectSettings' is set to true, but preferences file '...' could not be found!

This is caused by the tycho-compiler-plugin's setting
"useProjectSettings" that defaults to "true".

The documentation of the plugin states:

If set to true, the settings file
${project.basedir}/.settings/org.eclipse.jdt.core.prefs will be passed
to the compiler. If the file is not present, the build will not fail.

We do not use project specific JDT settings at all, so we can set the
option to false globally and get rid of that warning.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>