[homematic] Prevent thread leak on bridge handler disposal (#6375)
* [homematic] Prevent thread leak on bridge handler disposal

- the re-initialization of the handler is now cancelled when the handler
is disposed by the framework

When the initialization of the HomematicBridgeHandler fails, a
re-initialization is scheduled. This re-initialization was never stopped
when the handler was disposed by the framework. So in case the framework
disposes the handler while such a re-initialization was scheduled, the
initialization thread continuously tried to initialize the handler
despite it already being disposed.

Signed-off-by: Florian Stolte <fstolte@itemis.de>