rfc2217: set connect (not default) timeout
The "timeout" parameter of the our "open" method has uses a timeout
parameter.
That parameter defined the timeout we may block until the port is
opened.

The RFC2217 implementation sets the timeout parameter for the
unterlaying telnet client using that method "setDefaultTimeout".

Reading the code of the telnet client it seems the JavaDoc for that
method is misleading.

The default timeout is not used on opening the port but the connect
timeout is.

The default timeout will influence the reading.
If you try to read from the (virtual) serial port and there is no input
data for a period of "default timeout", an exception is thrown.

But that is not the "timeout" parameter on our open is used for.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>