Improve tests by mocking out Quartz scheduler
The previous testing strategy has been thrown away in favour of a mock version of the Quartz scheduler. The mock implements just enough Scheduler functionality to allow the code to work. This will hopefully make the tests 100% repeatable.

I have validated running this version of the test against the original TimerImpl / ScriptExecution, and as expected all tests pass apart from the one where it tries to reschedule the timer after it has terminated.

Signed-off-by: Jon Evans <jon.evans@pobox.com>