thing handler config update: don't initialize initialized handler
The current implementation of the thing handler callback triggers a
thing handler initialization if the "configuration updated" callback is
called.
If the handler is already initialized a debug log is written that an
already initialized handler should be initialized.
IMHO the message could lead to the intention we should not call the
configuration updated callback if the handler is initialized or we have
to un-initialize the handler in front of.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>