Extract ItemBuilderFactory from ItemRegistry (#6556)
The ItemRegistry did offer `newItemBuilder` methods which depended on a list of ItemFactories. In the ItemRegisryImpl nothing else did depend on the ItemFactories. With ESH in the Felix OSGi runtime there are circumstances where a circular dependency is detected for an ItemProvider implementation which depends on the ItemRegistry for ItemBuilder access.
This extracts the new `ItemBuilderFactory` service from the ItemRegistry and provides the same API. The now deprecated API from ItemRegistry provides a fallback implementation based on a single instance of CoreItemFactory.

This of cause API breaking.

Signed-off-by: Henning Treu <henning.treu@googlemail.com>