Made storage in OAuth2 required. (#6546)
Storage service was optional. However this causes problems because it could mean the storage is not yet available when the OAuth2 service is used. This causes the component using OAuth2 to fail with an exception, and thus would require all kind of extra logic to handle this implementation detail. The original idea for OAuth2 was it should also work when storage is not available, but in practice this makes it difficult to use. Because if it's not available the end user has to handle it. Like reauthorization because refresh token is not available. Also it looks like the implementation wasn't silently working when storage was not available it threw exceptions in such cases. To make the implementation work the storage service was made required.
Also cleanup of java doc, annotations and removed unused variables.

Signed-off-by: Hilbrand Bouwkamp <hilbrand@h72.nl>