Suggest inclusion of sunrise/sunset in Tips
I suggest adding a sentence in Tips showing how to get sunrise/sunset.  Users coming from other automation systems often want to trigger at sunrise or sunset, but nowhere on this page can you find them.  You need to look at the Wiki page to figure out that this is the start or end of daylight.  So I propose to add a sentence mentioning that the start/end of daylight can be used for sunrise/sunset.