Make min and max value for slider configurable
Like for the setpoint, the min and max value for a slider can differ based
on the underlying thing that is being controlled, so it shouldn't be
hardcoded and a user should be able to change these values.

The default values, if the parameters are not given, are 0 (minValue) and
100 (maxValue), like they're currently, so that users shouldn't see a
difference when they're not changing their configurations.

Fixes #6776

Signed-off-by: Florian <florian.schmidt.welzow@t-online.de>